requirements.txt
dateutil
Django #==1.11.2
Pillow #==4.1.1                                                                                                                                                                                  
django-cors-headers #==2.1.0                                                                                                                                                                     
djangorestframework #==3.6.3                                                                                                                                                                     
python-dateutil #==2.6.0                                                                                                                                                                         
typing #==3.6.1
 
